Projects: BSD: Fixed version checking
Fixed version checking in system.tcl. It was commented out, likely because this project just happens to also work in 2015.4. The fact that this project works in 2015.4 is an anomaly to the norm, which is that block diagram projects don't normally work between Vivado versions. The reason this project happens to work in both versions is because Xilinx did not update any of the IP cores that are used in it. The project will still not work with previous versions, nor with future versions (2016.X, etc.) so the version check has been added back in to throw an error when using any version but 2015.3 to generate the project